test(unit): move Cassandra Docker cluster tests
unit_tests/unit/test_cassandra_docker_cluster.py: move from unit_tests/test_cassandra_docker_cluster.py and centralize minimal cluster construction instead of patching __init__ in each test.
